refactor the db clients structure
